refactor: 예외 발생 시 사용자에게 보여질 행위를 담당하는 객체 생성
