+++ src/main/java/ladder/Application.java
@@ -1,10 +1,11 @@
 package ladder;
 
+import ladder.controller.FrontExceptionController;
 import ladder.controller.LadderGameController;
 
 public class Application {
 
     public static void main(String[] args) {
-        new LadderGameController().run();
+        new LadderGameController(new FrontExceptionController()).run();
     }
 }
